requirements.txt
streamlit>=1.28.0
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0
